requirements_unified.txt
# VoiceTree Unified Architecture Dependencies
# Core requirements for the new unified system

# LLM Integration
google-genai>=0.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration and Environment
python-dotenv>=1.0.0

# Async and Concurrency
asyncio
aiofiles>=23.0.0

# Data Validation and Serialization
typing-extensions>=4.8.0

# Logging and Monitoring
structlog>=23.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Optional: Enhanced features
# langchain>=0.1.0  # If using LangGraph features
# redis>=4.6.0      # If using Redis for caching
# prometheus-client>=0.19.0  # If using Prometheus metrics

# Legacy compatibility (temporary)
# These can be removed after migration is complete
google-generativeai  # Legacy Google AI SDK
langgraph  # Legacy workflow system 